Update soldiermre.lua
MRE now checks Primary instead of Alt, because otherwise it doesn't work with 2-handed grenades.

I have tried, and tried, and tried, and tried, and tried to wrestle with this script to get it to check alt first, THEN primary, but nothing works. So I'm doing it the crude way. If an actual lua coder has a better way of going about this, feel free to let me know.